Atomic bomb commit II: MRPT moves to Eigen! Status: - All unit tests work in MSVC. One fails in GCC, will need investigation (might be a bug in Eigen??) - All libs and apps build, and apparently work, but will need even more checking to be 100% sure. - Building of external programs via FIND_PACKAGE(MRPT) should work but will be modified to allow using non-embedded versions of Eigen.

+++ libs/base/include/mrpt/math/kmeans.h
@@ -86,7 +86,7 @@ namespace mrpt
 					}
 					else
 					{
-						ASSERTMSG_(dims==it->size(),"All points must have the same dimensionality.")
+						ASSERTMSG_(size_t(dims)==size_t(it->size()),"All points must have the same dimensionality.")
 					}
 
 					::memcpy(trg_ptr, &(*it)[0], dims*sizeof(typename TInnerVector::value_type));
