+++ libs/slam/src/slam/CMonteCarloLocalization3D.cpp
@@ -237,7 +237,7 @@ void CMonteCarloLocalization3D::PF_SLAM_implementation_replaceByNewParticleSet(
 	const vector_double		&newParticlesWeight,
 	const vector<size_t>	&newParticlesDerivedFromIdx )  const
 {
-	ASSERT_(newParticlesWeight.size()==newParticles.size())
+	ASSERT_(size_t(newParticlesWeight.size())==newParticles.size())
 
 	// ---------------------------------------------------------------------------------
 	// Substitute old by new particle set:
